refactor: capture tracer context in snapshot ops
